refactor: extract helpers resolving relation metadata into own files
